chore: Moved all logger related definitions to service/logger/definition/.
