+++ packages/architectura/src/definition/interface/_index.mts
@@ -7,8 +7,8 @@ export * from '../../core/server/definition/interface/https-server-configuration
 export * from '../../core/server/definition/interface/https-server-instantiation.interface.mjs';
 export * from '../../service/executor/definition/interface/executor-instantiation.interface.mjs';
 export * from '../../service/file-system/definition/interface/file-system-error.interface.mjs';
+export * from '../../service/logger/definition/interface/logger.interface.mjs';
 export * from './logger-service-write.interface.mjs';
-export * from './logger.interface.mjs';
 export * from './session-instantiation.interface.mjs';
 export * from './session.interface.mjs';
 export * from './stack-trace-line-parsing-result.interface.mjs';
