+++ packages/architectura/src/service/logger/definition/interface/logger-service-write.interface.mts
