+++ packages/architectura/src/service/logger/definition/interface/logger.interface.mts
