+++ packages/architectura/src/service/logger/logger.proxy.mts
@@ -1,7 +1,8 @@
 
+import type { LoggerInterface } from "./definition/interface/logger.interface.mjs";
+
 import type { LoggerService } from "./logger.service.mjs";
 
-import type { LoggerInterface } from "../../definition/interface/logger.interface.mjs";
 
 class LoggerProxy
 {
